src/main/java/com/vani/model/Student.java
package com.vani.model;

import org.springframework.data.annotation.Id;
import org.springframework.data.mongodb.core.mapping.Document;
@Document
public class Student {
 @Id
 private int ID;
 private String Name;
 private int age;
 private int salary;
 
 public Student() {}
 public Student(int ID, String name, int age, int salary) {
 super();
 this.ID = ID;
 Name = name;
 this.age = age;
 this.salary = salary;
 }
 public int getSalary() {
 return salary;
 }
 public void setSalary(int salary) {
 this.salary = salary;
 }
 public int getAge() {
 return age;
 }
 public void setAge(int age) {
 this.age = age;
 }
 public String getName() {
 return Name;
 }
 public void setName(String name) {
 Name = name;
 }
 public int getID() {
 return ID;
 }
 public void setID(int ID) {
 this.ID = ID;
 }
}

